Modified Gatling Loop
Modified gatling loop to avoid unfortunate face-killings. Expected
behavior: Trigger pressed - Start sequence. If not charged, charge and
set charged to true. Move to second frame. Check if trigger is still
pressed. If so, shoot and turn barrel. Things to test: If trigger is
bumped, the thing should charge and not shoot. Then, if trigger is
bumped, the thing should not charge and shoot.